+++ Source/equipment/lance-base.js
@@ -1,10 +1,10 @@
 const EquipmentTemplate = require('../../Classes/EquipmentTemplate.js');
 const { dealDamage, addModifier, getFullName } = require('../combatantDAO.js');
 
-module.exports = new EquipmentTemplate("Lance", "Strike a foe for @{damage} @{element} damage and gain @{mod1Stacks} @{mod1}", "Damage x@{critBonus}", "Earth", effect, ["Guarding Lance", "Reckless Lance", "Slowing Lance"])
+module.exports = new EquipmentTemplate("Lance", "Strike a foe for @{damage} @{element} damage (double increase from Power Up)", "Damage x@{critBonus}", "Earth", effect, ["Accelerating Lance", "Piercing Lance", "Vigilant Lance"])
 	.setCategory("Weapon")
 	.setTargetingTags({ target: "single", team: "enemy" })
-	.setModifiers([{ name: "Stagger", stacks: 1 }, { name: "Power Up", stacks: 25 }])
+	.setModifiers([{ name: "Stagger", stacks: 1 }])
 	.setCost(200)
 	.setUses(10)
 	.setDamage(75);
@@ -14,13 +14,15 @@ function effect([target], user, isCrit, adventure) {
 		return ` ${getFullName(target, adventure.room.enemyTitles)} was already dead!`;
 	}
 
-	let { element, modifiers: [elementStagger, powerUp], damage, critBonus } = module.exports;
+	let { element, modifiers: [elementStagger], damage, critBonus } = module.exports;
 	if (user.element === element) {
 		addModifier(target, elementStagger);
 	}
+	const powerUpStacks = user.getModifierStacks("Power Up");
+	damage += powerUpStacks;
 	if (isCrit) {
 		damage *= critBonus;
+		damage += powerUpStacks;
 	}
-	addModifier(user, powerUp);
 	return dealDamage(target, user, damage, false, element, adventure);
 }